requirements.txt
pyTigerGraph==1.6
neo4j-driver==5.3.0
apoc==0.12.0
langchain==0.2.5
mkdocs==1.6.0
mkdocs-material==9.5.27